pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "options-wheel"
version = "0.1.0"
description = "An automated options wheel trading strategy."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "python-dotenv",
    "pandas>=1.5",
    "numpy>=1.23",
    "pytz",
    "requests>=2.28",
    "alpaca-py",
    "finnhub-python",
    "rich>=14.0",
    "typer>=0.9.0",
    "ta>=0.11",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
run-strategy = "scripts.run_strategy:main"
run-screener = "scripts.run_screener:main"

[tool.setuptools.packages.find]
where = ["."]
exclude = ["archive"]
